requirements.txt
pydicom>=2.4.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
matplotlib>=3.7.0
pylibjpeg>=1.4.0
pylibjpeg-libjpeg>=1.3.0
python-gdcm>=3.0.20
torch>=2.1.0
torchvision>=0.16.0
scikit-learn>=1.3.0
pandas>=2.0.0
tqdm>=4.65.0
PyQt6>=6.5.0
PyQt6-Qt6>=6.5.0 